requirements.txt
litellm>=1.30.0
typer[all]>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
